fix: move applet api to src/entity/applet/api
