+++ src/entities/applet/api/index.ts
@@ -1,1 +1,2 @@
 export { useAppletListQuery } from "./useAppletListQuery"
+export { useAppletByIdQuery } from "./useAppletByIdQuery"
