+++ src/entities/applet/api/useAppletByIdQuery.ts
