+++ src/entities/applet/index.ts
@@ -1,3 +1,4 @@
 export * from "./ui"
 export * from "./lib"
+export * from "./api"
 export * as appletModel from "./model"
