+++ src/widgets/ActivityGroups/index.ts
@@ -1,2 +1,1 @@
 export * from "./ui"
-export * from "./api"
